[dino_catalog] File management   Remove unused file command_error.rb   Move utility classes to util/ directory
