--- dino_catalog/command_error.rb
@@ -1,1 +1,0 @@
-class CommandError < RuntimeError; end
