+++ dino_catalog/util/data_transformer.rb
