+++ dino_catalog/util/registry.rb
